big rewrite to C#14 & net10 support
* touched almost every single file - C#14 & nullability annotations are now standard (with no nullability-related warnings)
* removed almost all WinForms designer code, replaced with custom layout system (should be better for monitor DPI awareness, consistency and future development)
* new GDI multi-line formatted text layout engine (correct hyperlink rendering, almost certainly has issues with RTL languages in commit subject and body fields for now)
* removed GDI+ text renderer (too many issues, very much legacy, inferior rendering quality)
* dark theme support in dialogs
* graph visual configuration has a preview now
* AVX/SSE SIMD-accelerated hex hash string to binary parsers
* dark theme color palette changes (attempt to follow WinUI colors, expect higher contrast, not finished)
* GitLab plugin improvements (better access token management, show all pipelines for commit)
* TeamCity plugin improvements (moved to access token auth, show builds for selected commit)

+++ gitter.fw.prj/Options/WindowsExplorer.cs
@@ -38,24 +38,22 @@ static class WindowsExplorer
 			try
 			{
 				using var key = Registry.ClassesRoot.OpenSubKey($@"Directory\shell\{SubKeyName}\command", writable: false);
-				if(key is not null)
+				if(key is null) return false;
+				var value = (string?)key.GetValue(null, string.Empty);
+				if(string.IsNullOrEmpty(value)) return false;
+				if(value!.EndsWith(" \"%1\""))
 				{
-					var value = (string)key.GetValue(null, string.Empty);
-					if(value == string.Empty) return false;
-					if(value.EndsWith(" \"%1\""))
+					value = value.Substring(0, value.Length - 5);
+					if(value.StartsWith("\"") && value.EndsWith("\""))
 					{
-						value = value.Substring(0, value.Length - 5);
-						if(value.StartsWith("\"") && value.EndsWith("\""))
-						{
-							value = value.Substring(1, value.Length - 2);
-							value = Path.GetFullPath(value);
-							return value == GetAppPath();
-						}
+						value = value.Substring(1, value.Length - 2);
+						value = Path.GetFullPath(value);
+						return value == GetAppPath();
 					}
 				}
 				return false;
 			}
-			catch(Exception exc) when(!exc.IsCritical())
+			catch(Exception exc) when(!exc.IsCritical)
 			{
 				return false;
 			}
@@ -67,6 +65,7 @@ static class WindowsExplorer
 		try
 		{
 			using var key = Registry.ClassesRoot.OpenSubKey(@"Directory\shell", writable: true);
+			if(key is null) return;
 			using var gitterKey = key.CreateSubKey(SubKeyName, RegistryKeyPermissionCheck.ReadWriteSubTree, RegistryOptions.None);
 			gitterKey.SetValue(null, @"Open with gitter");
 			using var commandKey = gitterKey.CreateSubKey("command", RegistryKeyPermissionCheck.ReadWriteSubTree, RegistryOptions.None);
@@ -78,7 +77,7 @@ static class WindowsExplorer
 			appPath += " \"%1\"";
 			commandKey.SetValue(null, appPath);
 		}
-		catch(Exception exc) when(!exc.IsCritical())
+		catch(Exception exc) when(!exc.IsCritical)
 		{
 		}
 	}
@@ -88,9 +87,10 @@ static class WindowsExplorer
 		try
 		{
 			using var key = Registry.ClassesRoot.OpenSubKey(@"Directory\shell", writable: true);
+			if(key is null) return;
 			key.DeleteSubKeyTree(SubKeyName, throwOnMissingSubKey: false);
 		}
-		catch(Exception exc) when(!exc.IsCritical())
+		catch(Exception exc) when(!exc.IsCritical)
 		{
 		}
 	}
